+++ frontend/app/api/admin/diagnostics/meeting-counts/route.ts
@@ -2,7 +2,7 @@ import { Role } from "@prisma/client";
 import { NextResponse } from "next/server";
 import { requireRole } from "../../../../../services/auth";
 import { calendarIdForCategory } from "../../../../../services/googleCalendar";
-import { isDateSuspended } from "../../../../../util/meetingOccurrences";
+import { isDateSuspended } from "../../../../../util/meetings/meetingOccurrences";
 import { formatETDateString } from "../../../../../util/date/timeUtils";
 import { prisma } from "../../../../../lib/prisma";
