+++ frontend/app/api/admin/diagnostics/suspended/route.ts
@@ -1,7 +1,7 @@
 import { Role } from "@prisma/client";
 import { NextResponse } from "next/server";
 import { requireRole } from "../../../../../services/auth";
-import { getUnresolvedSuspension } from "../../../../../util/suspension";
+import { getUnresolvedSuspension } from "../../../../../util/meetings/suspension";
 import { formatETDateString } from "../../../../../util/date/timeUtils";
 import { prisma } from "../../../../../lib/prisma";
