+++ frontend/app/api/update/meeting/suspend/route.ts
@@ -3,8 +3,8 @@ import { NextResponse, after } from 'next/server';
 import { requireRole } from '../../../../../services/auth';
 import { trimCalendarEventSeries, deleteCalendarEvent, calendarIdsForMeeting } from '../../../../../services/googleCalendar';
 import { formatETDateString, getETDayBounds } from '../../../../../util/date/timeUtils';
-import { addOneETDay } from '../../../../../util/meetingOccurrences';
-import { getUnresolvedSuspension, reconcilePendingResume, createPendingResumeSeries, MeetingWithPattern } from '../../../../../util/suspension';
+import { addOneETDay } from '../../../../../util/meetings/meetingOccurrences';
+import { getUnresolvedSuspension, reconcilePendingResume, createPendingResumeSeries, MeetingWithPattern } from '../../../../../util/meetings/suspension';
 import { prisma } from '../../../../../lib/prisma';
 
 // MongoDB doesn't support SQL-style isolation levels (no Prisma isolationLevel option), and
