+++ frontend/tests/integration/meeting-occurrences-range.test.ts
@@ -1,6 +1,6 @@
 import { seedMeeting, seedRecurringMeeting } from "../factories/meeting";
 import { disconnectTestPrismaClient } from "../factories/db";
-import { getMeetingsForRange } from "../../util/meetingOccurrences";
+import { getMeetingsForRange } from "../../util/meetings/meetingOccurrences";
 import { convertETToUTC, addDaysToETDateString, formatETDateString } from "../../util/date/timeUtils";
 
 afterAll(async () => {
