+++ frontend/util/meetings/resourceOverlap.ts
@@ -1,6 +1,7 @@
+import "server-only";
 import { Prisma } from "@prisma/client";
-import { prisma } from "../lib/prisma";
-import { formatETDateString } from "./date/timeUtils";
+import { prisma } from "../../lib/prisma";
+import { formatETDateString } from "../date/timeUtils";
 import { matchesRecurrencePattern, adjustOccurrenceToDate, isDateSuspended } from "./meetingOccurrences";
 
 type SuspensionWindow = { from: Date; to: Date | null };
